requirements.txt
numpy==1.23.5
Pillow==9.5.0
streamlit==1.26.0
tensorflow==2.12.0